added --with-forces parameter
If --with-forces True is passed to train.py or play.py, touch sensor data
is included in the observation space. Currently, this is hardcoded and only works
with my gym fork.
Obviously, if play.py is used with forces, the given policy has to be
trained with forces as well.